# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Lynx Analytics — Plataforma para micro-SaaS brasileiros
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/revenue/timeline:
    get:
      operationId: getRevenueTimeline
      tags: [revenue]
      summary: Get MRR timeline for charting
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Revenue timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueTimeline"
  /projects/{projectId}/revenue/events:
    get:
      operationId: listRevenueEvents
      tags: [revenue]
      summary: List raw revenue events (payments, cancellations, upgrades)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [payment, cancellation, upgrade, downgrade, refund, new_subscription]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Revenue events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEventListResponse"
    post:
      operationId: createRevenueEvent
      tags: [revenue]
      summary: Manually record a revenue event
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRevenueEventRequest"
      responses:
        "201":
          description: Revenue event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueEvent"
  /projects/{projectId}/revenue/plans:
    get:
      operationId: getRevenuePlans
      tags: [revenue]
      summary: Revenue breakdown by plan
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revenue by plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanRevenue"

  # ─── CUSTOMERS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RevenueTimeline:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              mrr:
                type: number
              arr:
                type: number
              newMrr:
                type: number
              churnedMrr:
                type: number
              expansionMrr:
                type: number
            required: [date, mrr, arr, newMrr, churnedMrr, expansionMrr]
      required: [data]
    RevenueEvent:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [payment, cancellation, upgrade, downgrade, refund, new_subscription]
        amount:
          type: number
        currency:
          type: string
          default: BRL
        customerId:
          type: string
        customerEmail:
          type: string
        plan:
          type: string
        description:
          type: string
        externalId:
          type: string
        source:
          type: string
          enum: [abacatepay, manual, api]
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      required: [id, projectId, type, amount, currency, timestamp, source]
    RevenueEventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RevenueEvent"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [events, total, limit, offset]
    CreateRevenueEventRequest:
      type: object
      properties:
        type:
          type: string
          enum: [payment, cancellation, upgrade, downgrade, refund, new_subscription]
        amount:
          type: number
        currency:
          type: string
          default: BRL
        customerId:
          type: string
        customerEmail:
          type: string
        plan:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [type, amount]
    PlanRevenue:
      type: object
      properties:
        plan:
          type: string
        customers:
          type: integer
        mrr:
          type: number
        percentage:
          type: number
      required: [plan, customers, mrr, percentage]

    # ─── CUSTOMER SCHEMAS ───────────────────────────────────────────────────
...